+++ site/package.json
@@ -38,13 +38,9 @@
     "html-webpack-plugin": "^2.28.0",
     "json-loader": "^0.5.4",
     "markdown-doc-loader": "^1.0.0",
-    "postcss": "^5.1.2",
-    "postcss-easy-import": "^2.0.0",
     "postcss-loader": "^1.3.3",
-    "postcss-scss": "^0.4.1",
-    "precss": "^1.4.0",
     "react-hot-loader": "next",
-    "react-markdown-doc-loader": "^1.0.1",
+    "react-markdown-doc-loader": "^1.0.2",
     "style-loader": "^0.14.1",
     "url-loader": "^0.5.8",
     "webpack": "^2.2.1",
