+++ site/src/components/PageHeader/style.pcss
@@ -34,7 +34,7 @@
 
     &__item {
       float: left;
-      height: 50px;
+      height: 100%;
     }
 
     &__subnavs {
@@ -51,13 +51,13 @@
 
       a,
       span {
-        line-height: 50px;
+        line-height: $top-subnav-height;
       }
     }
 
     &__link {
       display: block;
-      line-height: 80px;
+      line-height: $top-nav-height;
       color: #333;
       font-size: 16px;
       margin: 0 20px;
